+++ mod-source-record-storage-server/src/main/java/org/folio/services/handlers/InstancePostProcessingEventHandler.java
@@ -1,42 +1,6 @@
 package org.folio.services.handlers;
 
-import static java.lang.String.format;
-import static org.apache.commons.lang.StringUtils.isNotEmpty;
-import static org.folio.dao.util.RecordDaoUtil.filterRecordByInstanceId;
-import static org.folio.dao.util.RecordDaoUtil.filterRecordByNotSnapshotId;
-import static org.folio.rest.jaxrs.model.EntityType.INSTANCE;
-import static org.folio.rest.jaxrs.model.EntityType.MARC_BIBLIOGRAPHIC;
-import static org.folio.rest.jaxrs.model.ProfileSnapshotWrapper.ContentType.MAPPING_PROFILE;
-import static org.folio.services.util.AdditionalFieldsUtil.TAG_999;
-import static org.folio.services.util.EventHandlingUtil.sendEventWithPayload;
-
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.concurrent.CompletableFuture;
-
 import com.fasterxml.jackson.databind.ObjectMapper;
-
-import org.apache.commons.lang.StringUtils;
-import org.apache.commons.lang3.tuple.Pair;
-import org.folio.DataImportEventPayload;
-import org.folio.MappingProfile;
-import org.folio.dao.RecordDao;
-import org.folio.dao.util.ParsedRecordDaoUtil;
-import org.folio.dao.util.RecordType;
-import org.folio.processing.events.services.handler.EventHandler;
-import org.folio.processing.exceptions.EventProcessingException;
-import org.folio.rest.jaxrs.model.AdditionalInfo;
-import org.folio.rest.jaxrs.model.EntityType;
-import org.folio.rest.jaxrs.model.ExternalIdsHolder;
-import org.folio.rest.jaxrs.model.Record;
-import org.folio.rest.util.OkapiConnectionParams;
-import org.folio.services.exceptions.PostProcessingException;
-import org.folio.services.util.AdditionalFieldsUtil;
-import org.jooq.Condition;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Component;
-
 import io.vertx.core.CompositeFuture;
 import io.vertx.core.Future;
 import io.vertx.core.Promise;
@@ -45,24 +9,66 @@ import io.vertx.core.json.Json;
 import io.vertx.core.json.JsonObject;
 import io.vertx.core.logging.Logger;
 import io.vertx.core.logging.LoggerFactory;
+import io.vertx.kafka.client.producer.KafkaHeader;
+import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.tuple.Pair;
+import org.folio.DataImportEventPayload;
+import org.folio.MappingProfile;
+import org.folio.dao.RecordDao;
+import org.folio.dao.util.ParsedRecordDaoUtil;
+import org.folio.dao.util.RecordType;
+import org.folio.kafka.KafkaConfig;
+import org.folio.processing.events.services.handler.EventHandler;
+import org.folio.processing.exceptions.EventProcessingException;
+import org.folio.rest.jaxrs.model.AdditionalInfo;
+import org.folio.rest.jaxrs.model.EntityType;
+import org.folio.rest.jaxrs.model.ExternalIdsHolder;
+import org.folio.rest.jaxrs.model.Record;
+import org.folio.services.exceptions.PostProcessingException;
+import org.folio.services.util.AdditionalFieldsUtil;
+import org.jooq.Condition;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Component;
+
+import java.util.ArrayList;
+import java.util.HashMap;
+import java.util.List;
+import java.util.concurrent.CompletableFuture;
+import java.util.concurrent.atomic.AtomicInteger;
+
+import static java.lang.String.format;
+import static org.apache.commons.lang.StringUtils.isNotEmpty;
+import static org.folio.dao.util.RecordDaoUtil.filterRecordByInstanceId;
+import static org.folio.dao.util.RecordDaoUtil.filterRecordByNotSnapshotId;
+import static org.folio.rest.jaxrs.model.DataImportEventTypes.DI_SRS_MARC_BIB_INSTANCE_HRID_SET;
+import static org.folio.rest.jaxrs.model.EntityType.INSTANCE;
+import static org.folio.rest.jaxrs.model.EntityType.MARC_BIBLIOGRAPHIC;
+import static org.folio.rest.jaxrs.model.ProfileSnapshotWrapper.ContentType.MAPPING_PROFILE;
+import static org.folio.rest.util.OkapiConnectionParams.OKAPI_TENANT_HEADER;
+import static org.folio.rest.util.OkapiConnectionParams.OKAPI_TOKEN_HEADER;
+import static org.folio.rest.util.OkapiConnectionParams.OKAPI_URL_HEADER;
+import static org.folio.services.util.AdditionalFieldsUtil.TAG_999;
+import static org.folio.services.util.EventHandlingUtil.sendEventToKafka;
 
 @Component
 public class InstancePostProcessingEventHandler implements EventHandler {
 
   private static final Logger LOG = LoggerFactory.getLogger(InstancePostProcessingEventHandler.class);
+  private static final AtomicInteger indexer = new AtomicInteger();
 
   private static final String FAIL_MSG = "Failed to handle instance event {}";
   private static final String EVENT_HAS_NO_DATA_MSG = "Failed to handle Instance event, cause event payload context does not contain INSTANCE and/or MARC_BIBLIOGRAPHIC data";
-  private static final String RECORD_UPDATED_EVENT_TYPE = "DI_SRS_MARC_BIB_INSTANCE_HRID_SET";
   private static final String DATA_IMPORT_IDENTIFIER = "DI";
 
   private final RecordDao recordDao;
   private final Vertx vertx;
+  private final KafkaConfig kafkaConfig;
 
   @Autowired
-  public InstancePostProcessingEventHandler(final RecordDao recordDao, Vertx vertx) {
+  public InstancePostProcessingEventHandler(final RecordDao recordDao, Vertx vertx, KafkaConfig kafkaConfig) {
     this.recordDao = recordDao;
     this.vertx = vertx;
+    this.kafkaConfig = kafkaConfig;
   }
 
   /**
@@ -98,10 +104,11 @@ public class InstancePostProcessingEventHandler implements EventHandler {
             HashMap<String, String> context = dataImportEventPayload.getContext();
             context.put(Record.RecordType.MARC.value(), Json.encode(record));
             context.put(DATA_IMPORT_IDENTIFIER, "true");
-            OkapiConnectionParams params = getConnectionParams(dataImportEventPayload);
-            sendEventWithPayload(Json.encode(context), RECORD_UPDATED_EVENT_TYPE, params);
-
+            List<KafkaHeader> kafkaHeaders = getKafkaHeaders(dataImportEventPayload);
+            String key = String.valueOf(indexer.incrementAndGet() % 100);
             context.put(MARC_BIBLIOGRAPHIC.value(), Json.encode(record));
+            sendEventToKafka(dataImportEventPayload.getTenant(), Json.encode(context), DI_SRS_MARC_BIB_INSTANCE_HRID_SET.value(),
+              kafkaHeaders, kafkaConfig, key);
             future.complete(dataImportEventPayload);
           } else {
             LOG.error(FAIL_MSG, updateAr.cause());
@@ -124,13 +131,11 @@ public class InstancePostProcessingEventHandler implements EventHandler {
     }
   }
 
-  private OkapiConnectionParams getConnectionParams(DataImportEventPayload dataImportEventPayload) {
-    OkapiConnectionParams params = new OkapiConnectionParams();
-    params.setOkapiUrl(dataImportEventPayload.getOkapiUrl());
-    params.setTenantId(dataImportEventPayload.getTenant());
-    params.setToken(dataImportEventPayload.getToken());
-    params.setVertx(vertx);
-    return params;
+  private List<KafkaHeader> getKafkaHeaders(DataImportEventPayload eventPayload) {
+    return List.of(
+      KafkaHeader.header(OKAPI_URL_HEADER, eventPayload.getOkapiUrl()),
+      KafkaHeader.header(OKAPI_TENANT_HEADER, eventPayload.getTenant()),
+      KafkaHeader.header(OKAPI_TOKEN_HEADER, eventPayload.getToken()));
   }
 
   private Future<Void> updatePreviousRecords(String instanceId, String snapshotId, String tenantId) {