+++ mod-source-record-storage-server/src/test/java/org/folio/services/AbstractLBServiceTest.java
@@ -1,6 +1,8 @@
 package org.folio.services;
 
+import net.mguenther.kafka.junit.EmbeddedKafkaCluster;
 import org.folio.dao.PostgresClientFactory;
+import org.folio.kafka.KafkaConfig;
 import org.folio.rest.RestVerticle;
 import org.folio.rest.client.TenantClient;
 import org.folio.rest.jaxrs.model.Metadata;
@@ -17,8 +19,18 @@ import io.vertx.core.Vertx;
 import io.vertx.core.json.JsonObject;
 import io.vertx.ext.unit.Async;
 import io.vertx.ext.unit.TestContext;
+import org.junit.ClassRule;
+
+import static net.mguenther.kafka.junit.EmbeddedKafkaCluster.provisionWith;
+import static net.mguenther.kafka.junit.EmbeddedKafkaClusterConfig.useDefaults;
 
 public abstract class AbstractLBServiceTest {
+
+  private static final String KAFKA_HOST = "FOLIO_KAFKA_HOST";
+  private static final String KAFKA_PORT = "FOLIO_KAFKA_PORT";
+  private static final String OKAPI_URL_ENV = "OKAPI_URL";
+  private static final int PORT = NetworkUtils.nextFreePort();
+  protected static final String OKAPI_URL = "http://localhost:" + PORT;
 
   private static PostgreSQLContainer<?> postgresSQLContainer;
 
@@ -26,15 +38,30 @@ public abstract class AbstractLBServiceTest {
   static final String PARSED_MARC_RECORD_CONTENT_SAMPLE_PATH = "src/test/resources/parsedMarcRecordContent.sample";
 
   static final String TENANT_ID = "diku";
+  static final String TOKEN = "dummy";
 
   static PostgresClientFactory postgresClientFactory;
 
   static Vertx vertx;
+  static KafkaConfig kafkaConfig;
+
+  @ClassRule
+  public static EmbeddedKafkaCluster cluster = provisionWith(useDefaults());
 
   @BeforeClass
   public static void setUpClass(TestContext context) throws Exception {
     Async async = context.async();
     vertx = Vertx.vertx();
+
+    String[] hostAndPort = cluster.getBrokerList().split(":");
+    System.setProperty(KAFKA_HOST, hostAndPort[0]);
+    System.setProperty(KAFKA_PORT, hostAndPort[1]);
+    System.setProperty(OKAPI_URL_ENV, OKAPI_URL);
+
+    KafkaConfig kafkaConfig = KafkaConfig.builder()
+      .kafkaHost(hostAndPort[0])
+      .kafkaPort(hostAndPort[1])
+      .build();
 
     String postgresImage = PomReader.INSTANCE.getProps().getProperty("postgres.image");
     postgresSQLContainer = new PostgreSQLContainer<>(postgresImage);
@@ -61,6 +88,7 @@ public abstract class AbstractLBServiceTest {
         });
       } catch (Exception e) {
         e.printStackTrace();
+        async.complete();
       }
     });
   }
@@ -81,5 +109,5 @@ public abstract class AbstractLBServiceTest {
     context.assertEquals(expected.getUpdatedByUserId(), actual.getUpdatedByUserId());
     context.assertNotNull(actual.getUpdatedDate());
   }
-  
+
 }